requirement.txt
fastapi
python - dotenv
pydantic[email]
pyjwt
pymongo
motor
black
isort
bcrypt
